\name{RpsiFinder_example}
\alias{RpsiFinder_example}
\title{Get RpsiFinder example result}
\usage{RpsiFinder_example <- function()}
\description{This function will get an example psiFinder result from ePSI-seq sequencing data.}
\examples{RpsiFinder_example()#when finished, view your global environment, and you will detect Data named 'RpsiFinder_example'}
\value{tab-delimited text files include stop read information for each CMC-treatment and CMC-control group, e.g. treatStopRatio, ctrlStopRatio}
