\name{user_defined}
\alias{user_defined}
\title{Identify pseudouridylation sites based on user-defined thresholds}
\usage{
user_defined <- function(rocfile,
                        rRNAfile,
                        filtfile,
                        output_dir,
                        output_name)
}
\arguments{
  \item{rocfile}{User-defined input file of single sites information (normally with suffix _roc_plot.txt) generated from ground_truth() function.}
  \item{rRNAfile}{The bed file contains knonw rRNA pseudouridylation sites, default we recommend hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed.}
  \item{filtfile}{File to be filted.}
  \item{output_dir}{Directory of the output file.}
  \item{output_name}{Name of the output file.}
}
\description{This function will accept result file generated by ground_truth() function and give out evaluation on input ground truth rRNA pseudouridylation data set.}
\examples{
RpsiFinder_example()#when finished, view your global environment, and you will detect Data named 'RpsiFinder_example'

prefilt_out <- prefilt(RpsiFinder_res=RpsiFinder_example,
                        target_chrom="^chr[0-9|a-z|A-Z]*$",
                        target_base="T",
                        treatStopNum_thres=10,
                        output_dir="/home/chenzhr23/test_out/RpsiFinder_out",
                        output_file="test_prefilt.txt")

udfilt_out<-udfilt(RpsiFinder_res_file="/home/chenzhr23/test_out/RpsiFinder_out/test_prefilt.txt",
                    treatPreRpmFold_thres=1.5,
                    treatAfterRpmFold_thres=1.5,
                    treatStopRatio_thres=0.04,
                    preRpmFoldRatio_thres=1.5,
                    afterRpmFoldRatio_thres=1.5,
                    stopRatioFC_thres=1.5,
                    output_dir="/home/chenzhr23/test_out/RpsiFinder_out",
                    output_file="test_udfilt.txt")

ground_truth(RpsiFinder_res_file="/home/chenzhr23/test_out/RpsiFinder_out/test_udfilt.txt",
              known_rRNA_bed="../test_data/ground_truth/hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed",
              rrna_chr21_bed="../test_data/ground_truth/rrna_chr21.bed",
              output_dir="/home/chenzhr23/test_out/ground_truth",
              output_name="ground_truth")

user_defined(rocfile="/home/chenzhr23/test_out/ground_truth/ground_truth_roc_plot.txt",
              rRNAfile="../test_data/ground_truth/hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed",
              filtfile="../test_out/RpsiFinder_out/test_udfilt.txt",
              output_dir="/home/chenzhr23/test_out/user_defined",
              output_name="user_defined")

}
\value{Evaluation result (figures and tables) contain TPR, TNR, etc.}
